refactor: Moved hooks out of `colorizer` and into their own directory
