--- src/colorizer/utils/hooks/index.ts
@@ -1,1 +1,0 @@
-export { useShortcutKey } from "./useShortcutKey";
