+++ tests/setup.ts
@@ -1,8 +1,9 @@
 import "@testing-library/jest-dom/vitest";
+import "vitest-canvas-mock";
+
 import { cleanup } from "@testing-library/react";
 import { generateUUID } from "three/src/math/MathUtils";
 import { afterEach, vi } from "vitest";
-import "vitest-canvas-mock";
 
 // Fix for the following error:
 // `TypeError: The "obj" argument must be an instance of Blob. Received an instance of Blob`
